# perf: Use split op to improve LSTM performance