Before:

test_sequence_cell[cpu-forward-lstm] │     3,194,966ns 
test_sequence_cell[cpu-backward-lstm] │   331,443,152ns

---
After:

test_sequence_cell[cpu-forward-lstm] │   185,876ns (~17x speedup)
test_sequence_cell[cpu-backward-lstm] │ 1,203,305ns (~27x speedup)